Implement stock patching
